Change `salishsea prepare` to copy run description files instead of symlinking.
Run description YAML file, iodef.xml file, and xmlio_server.def file are
all now copied to run directory so that they accurately reflect what is
intended to be used in a queued run.
A side-effect of this is that the iodef.xml and xmlio_server.def files
are now included in the gathered run results, providing a more complete
record of how the run was conducted.