Add seawolf[123] to SYSTEM keyed dicts in run.py
seawolf[123] are the names of the orcinus login nodes. The value of SYSTEM gets
set to one of those when `salishsea run` is executed via an sftp client in the
SalishSeaNowcast run_NEMO_agrif worker. That was causing orcinus-specific
configuration in the run script to be omitted for those runs, in particular,
partition=QDR.