Implement stdout & stderr redirection for `salish`
This commit adds redirection of stdout and stderr to files within
`${RESULTS_DIR}` for the `salish` system, ensuring logging of output and error
streams. New tests were added to verify the functionality for both the
`--deflate` and non-deflate cases. Necessary changes were made in the
`_execute()` function and corresponding test cases to support this redirection.